Update Layout
Dunno if this will work, when I tried to load the one that's there now it failed.

I changed the left and right arrows to home and end when the function key is pressed. They were previously hue up and down. I moved hue up and down to the right side extra keys, the ones just above the right key. Swapped with home and end. I don't even use the backlight right now so I think this will be better.